pyproject.toml
[project]
name = "quizling"
version = "0.1.0"
description = "Generate multiple choice questions from documents using Azure OpenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
